ssl_options {mongolite}	R Documentation

Connection SSL options

Description

Set SSL options to connect to the MongoDB server.

Usage

ssl_options(cert = NULL, key = cert, ca = NULL, ca_dir = NULL,
  crl_file = NULL, allow_invalid_hostname = NULL,
  weak_cert_validation = NULL)

Arguments

cert
path to PEM file with client certificate, or a certificate as returned by openssl::read_cert()

key
path to PEM file with private key from the above certificate, or a key as returned by openssl::read_key(). This can be the same PEM file as cert.

ca
a certificate authority PEM file

ca_dir
directory with CA files

crl_file
file with revocations

allow_invalid_hostname
do not verify hostname on server certificate

weak_cert_validation
disable certificate verification

[Package mongolite version 1.6 Index]